Rename: ImageInput 경로 수정
