+++ components/DashBoard/CreateTodoForm.tsx
@@ -10,7 +10,7 @@ import TitleInput from "./inputs/TitleInput";
 import DescriptionInput from "./inputs/DescriptionInput";
 import DateInput from "./inputs/DateInput";
 import TagInput from "./inputs/TagInput";
-import ImageInput from "./ImageInput";
+import ImageInput from "./inputs/ImageInput";
 import UserInput from "./inputs/UserInput";
 
 const CreateTodoForm = ({ columnId, onClose, dashboardId }: TodoModalProps) => {
